refactoring exception handling

+++ backend/src/main/java/com/dashboard/service/EmployeeService.java
@@ -1,8 +1,12 @@
 package com.dashboard.service;
 
 import com.dashboard.exception.BadRequestException;
+import com.dashboard.exception.EmployeeNotFoundException;
+import com.dashboard.exception.TaskNotFoundException;
 import com.dashboard.model.Employee;
+import com.dashboard.model.Task;
 import com.dashboard.repository.EmployeeRepository;
+import com.dashboard.repository.TaskRepository;
 import com.dashboard.service.interfaces.EmployeeServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,6 +19,10 @@ public class EmployeeService implements EmployeeServiceInterface {
 
     @Autowired
     private EmployeeRepository employeeRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
+
 
     @Override
     public void deleteAllEmployees(){
@@ -58,15 +66,47 @@ public class EmployeeService implements EmployeeServiceInterface {
     }
 
     @Override
-    public void deleteEmployeeById(Integer id){
-        employeeRepository.deleteById(id);
+    public void deleteEmployeeById(Integer employeeID){
+        final Optional<Employee> employee = employeeRepository.findById(employeeID);
+        if(!employee.isPresent()) throw new EmployeeNotFoundException();
+        employeeRepository.deleteById(employeeID);
     }
 
     @Override
     public void updateEmployee(Integer employeeId, Employee employee) {
+        final Optional<Employee> emp = employeeRepository.findById(employeeId);
+        if(!emp.isPresent()) throw new EmployeeNotFoundException();
         employee.setId(employeeId);
         employeeRepository.deleteById(employeeId);
         employeeRepository.save(employee);
+    }
+
+    @Override
+    public void assignTask(Integer employeeId, Integer taskID) {
+
+        final Optional<Task> task = taskRepository.findById(taskID);
+
+        if(!task.isPresent()) throw new TaskNotFoundException();
+
+        final Optional<Employee> employee = employeeRepository.findById(employeeId);
+
+        if(!employee.isPresent()) throw new EmployeeNotFoundException();
+
+        employee.get().getTasks().add(task.get());
+    }
+
+    @Override
+    public void unassignTask(Integer employeeId, Integer taskID) {
+
+        final Optional<Task> task = taskRepository.findById(taskID);
+
+        if(!task.isPresent()) throw new TaskNotFoundException();
+
+        final Optional<Employee> employee = employeeRepository.findById(employeeId);
+
+        if(!employee.isPresent()) throw new EmployeeNotFoundException();
+
+        employee.get().getTasks().remove(task.get());
     }
 
 }
